LPS-63470 update imports, move constants, fix compile
